fix issue saving depends_on steps in the UI, add validation for name
Providing an id as a string caused the mapping to fail and an error to propagate through, when no errors should have occurred. 

Also added a validate method to the step class as it seems to have allowed saving an empty named step, which means it won't be editable via the UI (no name = no link).