fix: issue if exception is thrown and engine is not defined to log it
This will use the engine's log method if engine is available, otherwise it will spit out an mtrace message. 

This might be desirable for now, but also it might be more desirable to update ensure engine's constructor can never throw an exception, and the dataflow is set using another method, which will make for a more consistent logging experience (e.g. it would logs to the appropriate instance's history also)