Issue #204: Cleanup step type classes.
Refactor to have stepdef and enginestep members set in the constructor of base_step.
Remove stepdef and enginestep from parameter list of all functions in step type classes.
Refactor step::get_steptype() to pass itself to the constuctor of the steptype class.
Refactor engine to pass the stepdef and itself to the constructor of the steptype, in order to generate the engine.
Change get_engine_step() to just return the engien step.
Add generate_engine_step() to generate to engine step.
Refactor all step type classes to call generate_engien_step instead of get_engine_step.