fix: calling abort multiple times won't cause an exception
Due to the exception being thrown when "something" aborts the dataflow, whether intentionally or by accident, the dataflow finish time was not being recorded, and other various cleanup activities potentially missed.

This should help to keep a nice log and resolve the issue with the finish time not being set.

Resolves #681